requirements.txt
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
kafka-python-ng==2.2.3
apache-airflow-providers-apache-kafka
apache-airflow-providers-postgres==5.0.0
airflow-operators
SQLAlchemy==1.4.53
ujson==5.9.0
python-dateutil==2.8.2
pytz==2023.4
colorama==0.4.6
pywin32==306; sys_platform == "win32"
requests==2.31.0
urllib3==2.0.7
pytest==7.4.4
pytest-mock==3.12.0
rich==13.7.0